cluster-ui: memoize columns on db pages, cleanup class components
Resolves: #104873

This change removes column computations from class component state and
instead memoizes them. Some cleanup as well to move some table column
logic outside the class component.

Release note: None
